requirements.txt
# Intent-Routed Agent POC - All Dependencies
# Install: pip install -r requirements.txt

# ============================================================================
# Core AI/ML Framework
# ============================================================================

# LangChain & LangGraph for agent workflows
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-experimental>=0.0.50
langchain-text-splitters>=0.0.1
langgraph>=0.0.25

# OpenAI for LLM and embeddings
openai>=1.10.0

# ============================================================================
# Vector Search & RAG
# ============================================================================

# FAISS for semantic search
faiss-cpu>=1.7.4
numpy>=1.24.0

# BM25 for keyword search
rank-bm25>=0.2.2

# ============================================================================
# API Service
# ============================================================================

# FastAPI web framework
fastapi>=0.108.0

# ASGI server
uvicorn[standard]>=0.25.0

# Data validation
pydantic>=2.5.0

# ============================================================================
# Web UI (Streamlit)
# ============================================================================

# Streamlit for interactive dashboard
streamlit>=1.28.0

# Data manipulation and visualization
pandas>=2.0.0
plotly>=5.17.0

# ============================================================================
# Utilities
# ============================================================================

# HTTP client
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Type annotations
typing-extensions>=4.9.0

# ============================================================================
# Optional: Observability (Uncomment if needed)
# ============================================================================

# LangSmith tracing (just set LANGCHAIN_TRACING_V2=true in .env)
# langsmith>=0.0.80

# OpenTelemetry (for advanced tracing)
# opentelemetry-api>=1.21.0
# opentelemetry-sdk>=1.21.0
# opentelemetry-exporter-otlp>=1.21.0
